model-context-protocol/pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.51.0",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
]
